+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- *(sandbox)* bound concurrent script workers and fail fast when the limit is reached ([#29](https://github.com/geeknonerd/stuntdouble/issues/29))
+
 ## [0.2.1](https://github.com/geeknonerd/stuntdouble/compare/v0.2.0...v0.2.1) - 2026-09-23
 
 ### Other